Fix local tests
