+++ realm/kotlin-extensions/build.gradle
@@ -61,6 +61,10 @@ android {
                 '../realm-library/src/syncTestUtils/java',
         ]
     }
+    compileOptions {
+        targetCompatibility 1.8
+        sourceCompatibility 1.8
+    }
 }
 
 dependencies {
