+++ realm/realm-library/src/syncIntegrationTest/java/io/realm/StandardIntegrationTest.java
@@ -44,20 +44,20 @@ public abstract class StandardIntegrationTest extends BaseIntegrationTest {
 
     @Before
     public void setupTest() throws IOException {
-        prepareEnvironmentForTest();
+        SyncTestUtils.prepareEnvironmentForTest();
     }
 
     @After
     public void teardownTest() {
         if (!looperThread.isRuleUsed() || looperThread.isTestComplete()) {
             // Non-looper tests can reset here
-            restoreEnvironmentAfterTest();
+            SyncTestUtils.restoreEnvironmentAfterTest();
         } else {
             // Otherwise we need to wait for the test to complete
             looperThread.runAfterTest(new Runnable() {
                 @Override
                 public void run() {
-                    restoreEnvironmentAfterTest();
+                    SyncTestUtils.restoreEnvironmentAfterTest();
                 }
             });
         }
